LayoutTests/canvas/philip/tests/2d.path.isPointInPath.nonfinite-expected.txt

< [index] >
2d.path.isPointInPath.nonfinite
isPointInPath() returns false for non-finite arguments
References: 2d.path.isPointInPath.nonfinite
Actual output:
Passed
